code clean up

+++ coderadar-graph/src/main/java/io/reflectoring/coderadar/graph/projectadministration/analyzerconfig/service/DeleteAnalyzerConfigurationAdapter.java
@@ -3,24 +3,22 @@ package io.reflectoring.coderadar.graph.projectadministration.analyzerconfig.ser
 import io.reflectoring.coderadar.graph.projectadministration.analyzerconfig.repository.AnalyzerConfigurationRepository;
 import io.reflectoring.coderadar.projectadministration.AnalyzerConfigurationNotFoundException;
 import io.reflectoring.coderadar.projectadministration.port.driven.analyzerconfig.DeleteAnalyzerConfigurationPort;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 public class DeleteAnalyzerConfigurationAdapter implements DeleteAnalyzerConfigurationPort {
   private final AnalyzerConfigurationRepository analyzerConfigurationRepository;
 
-  @Autowired
   public DeleteAnalyzerConfigurationAdapter(
       AnalyzerConfigurationRepository analyzerConfigurationRepository) {
     this.analyzerConfigurationRepository = analyzerConfigurationRepository;
   }
 
   @Override
-  public void deleteAnalyzerConfiguration(Long id) throws AnalyzerConfigurationNotFoundException {
-    analyzerConfigurationRepository
-        .findById(id)
-        .orElseThrow(() -> new AnalyzerConfigurationNotFoundException(id));
+  public void deleteAnalyzerConfiguration(Long id) {
+    if (!analyzerConfigurationRepository.existsById(id)) {
+      throw new AnalyzerConfigurationNotFoundException(id);
+    }
     analyzerConfigurationRepository.deleteById(id);
   }
 }
